cleanup + requested changes. #55
