+++ explorer/app/transformers/hca.ts
@@ -5,23 +5,10 @@
 import {
   ProjectListResponse,
   ProjectResponse,
-  FileResponse,
-  SampleResponse,
   FileListResponse,
   SampleListResponse,
 } from "../models/responses";
-import { DetailModel, ListViewModel } from "../models/viewModels";
-
-/**
- * Transforms a response to the projects endpoint into a viewModel, that will be used by
- * @see DetailContainer
- * @param value Api's response type
- * @returns @see DetailModel
- */
-export const projectDetailToView = (value: ProjectResponse): DetailModel => ({
-  json: JSON.stringify(value, null, 2),
-  detailName: value?.projects?.[0]?.projectTitle ?? "",
-});
+import { ListViewModel } from "../models/viewModels";
 
 /**
  * Function to get the id of the given detail type
@@ -30,28 +17,6 @@ export const projectDetailToView = (value: ProjectResponse): DetailModel => ({
  */
 export const getProjectId = (value: ProjectResponse) =>
   value.projects[0].projectId;
-
-/**
- * Transforms a response to the file endpoint into a viewModel, that will be used by
- * @see DetailContainer
- * @param value Api's response type
- * @returns @see DetailModel
- */
-export const fileDetailToView = (value: FileResponse): DetailModel => ({
-  json: JSON.stringify(value, null, 2),
-  detailName: value.files[0].name,
-});
-
-/**
- * Transforms a response to the samples endpoint into a viewModel, that will be used by
- * @see DetailContainer
- * @param value Api's response type
- * @returns @see DetailModel
- */
-export const sampleDetailToView = (value: SampleResponse): DetailModel => ({
-  json: JSON.stringify(value, null, 2),
-  detailName: value.samples[0].id,
-});
 
 /**
  * Transforms a list of project items to a viewModel for hca project
