+++ explorer/package.json
@@ -30,7 +30,8 @@
     "next-optimized-images": "^2.6.2",
     "react": "17.0.2",
     "react-dom": "17.0.2",
-    "react-table": "^7.7.0"
+    "react-table": "^7.7.0",
+    "uuid": "^8.3.2"
   },
   "devDependencies": {
     "@babel/core": "^7.17.10",
@@ -45,6 +46,7 @@
     "@types/react": "18.0.9",
     "@types/react-dom": "18.0.3",
     "@types/react-table": "^7.7.12",
+    "@types/uuid": "^8.3.4",
     "@typescript-eslint/eslint-plugin": "5.23.0",
     "babel-loader": "^8.2.5",
     "eslint": "8.15.0",
