+++ explorer/site-config/hca-dcp/dev/config.ts
@@ -2,12 +2,9 @@
 import { ELEMENT_ALIGNMENT } from "../../../app/common/entities";
 import { SiteConfig } from "../../../app/config/model";
 import {
-  fileDetailToView,
   filesListToView,
   getProjectId,
-  projectDetailToView,
   projectListToView,
-  sampleDetailToView,
   samplesListToView,
 } from "app/transformers/hca";
 import { sideColumn } from "./sideColumn";
@@ -28,7 +25,6 @@ const config: SiteConfig = {
       apiPath: "index/projects",
       route: "projects",
       listTransformer: projectListToView,
-      detailTransformer: projectDetailToView,
       getId: getProjectId,
       staticLoad: true,
     },
@@ -37,14 +33,12 @@ const config: SiteConfig = {
       apiPath: "index/files",
       route: "files",
       listTransformer: filesListToView,
-      detailTransformer: fileDetailToView,
     },
     {
       label: "Samples",
       apiPath: "index/samples",
       route: "samples",
       listTransformer: samplesListToView,
-      detailTransformer: sampleDetailToView,
     },
   ],
   layout: {
